feat(dna): update DnaDerivedData mixin for proper serialization
- Added DnaDerivedDataMixin interface to handle nFraction and nRunsCapped fields.
- Updated ObjectMapper configuration to use the new mixin for serialization.
- Renamed columns in dataset to ensure correct casing during serialization.
