+++ gbif/ingestion/spark-jobs/src/main/java/org/gbif/pipelines/spark/OccurrenceInterpretationPipeline.java
@@ -46,7 +46,6 @@ import org.gbif.pipelines.core.functions.SerializableSupplier;
 import org.gbif.pipelines.core.interpreters.metadata.MetadataInterpreter;
 import org.gbif.pipelines.core.pojo.HdfsConfigs;
 import org.gbif.pipelines.core.utils.FsUtils;
-import org.gbif.pipelines.core.utils.SerDeUtils;
 import org.gbif.pipelines.core.ws.metadata.MetadataServiceClient;
 import org.gbif.pipelines.io.avro.*;
 import org.gbif.pipelines.io.avro.grscicoll.GrscicollRecord;
@@ -74,9 +73,7 @@ import scala.Tuple2;
 public class OccurrenceInterpretationPipeline {
 
   static final ObjectMapper MAPPER =
-      new ObjectMapper()
-          .setSerializationInclusion(JsonInclude.Include.NON_EMPTY)
-          .addMixIn(DnaDerivedData.class, SerDeUtils.DnaDerivedDataMixin.class);
+      new ObjectMapper().setSerializationInclusion(JsonInclude.Include.NON_EMPTY);
 
   public static final String METRICS_FILENAME = "verbatim-to-occurrence.yml";
 
@@ -805,6 +802,12 @@ public class OccurrenceInterpretationPipeline {
     // for small datasets, to reduce the number of small files created, we coalesce to a single
     // shard
     dataset = dataset.coalesce(numshards);
+
+    // hack to serialize these 2 fields with the proper casing
+    dataset
+        .withColumnRenamed("NFraction", "nFraction")
+        .withColumnRenamed("NRunsCapped", "nRunsCapped");
+
     return dataset;
   }
 }
